cleaned up a little
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@3185 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/subpane/wb/wbuploader/UploadTableTree.java
@@ -448,19 +448,23 @@ public class UploadTableTree extends UploadTable
             for (Treeable defParent : defaultParents)
             {
                 Object key = ((DataModelObjBase)defParent).getId();
-                try
+                if (key != null)
                 {
-                    q.setParameter("theKey", key);
-                    DataModelObjBase obj = (DataModelObjBase)q.uniqueResult();
-                    session.beginTransaction();
-                    session.delete(obj);
-                    session.commit();
-                }
-                catch (Exception ex)
-                {
-                    //the delete may fail if another user has used or deleted uploaded records...
-                    //or if another UploadTreeTable has deleted the parent.
-                    log.info(ex);
+                    try
+                    {
+                        q.setParameter("theKey", key);
+                        DataModelObjBase obj = (DataModelObjBase) q.uniqueResult();
+                        session.beginTransaction();
+                        session.delete(obj);
+                        session.commit();
+                    }
+                    catch (Exception ex)
+                    {
+                        // the delete may fail if another user has used or deleted uploaded
+                        // records...
+                        // or if another UploadTreeTable has deleted the parent.
+                        log.info(ex);
+                    }
                 }
             }
         }
@@ -575,8 +579,6 @@ public class UploadTableTree extends UploadTable
             }
         }
         return false;
-        
-        //return fld.isRequired() && getTreeDefItem().getIsEnforced() != null && getTreeDefItem().getIsEnforced() && (fld.getValue() == null || fld.getValue().trim().equals(""));
     }
 
     /**
